Envelopa o InvitationService em um module
Para melhor organização do código com adições futuras.
